Move signals into sub-module
